Silme işlemini tamamladım. Güncelleme işlemini yapıyorum

+++ app/src/main/java/com/aek/notes/view/MainActivity.java
@@ -10,7 +10,7 @@ import android.view.MenuItem;
 import android.view.Window;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
+import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.SearchView;
 import androidx.fragment.app.FragmentTransaction;
@@ -19,12 +19,10 @@ import com.aek.notes.R;
 import com.aek.notes.core.constants.AppConstants;
 import com.aek.notes.core.util.AppUtils;
 import com.aek.notes.databinding.ActivityMainBinding;
-import com.aek.notes.model.ModelNote;
 import com.aek.notes.view.fragment.NoteListFragment;
 import com.aek.notes.viewmodel.ViewModelNote;
 import com.aek.notes.viewmodel.ViewModelNoteForm;
 import com.google.android.material.transition.platform.MaterialContainerTransformSharedElementCallback;
-import com.google.gson.Gson;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -58,11 +56,6 @@ public class MainActivity extends AppCompatActivity {
 
     }
 
-    @Override
-    protected void onResume() {
-        ViewModelNote.getInstance().loadLocalData(this);
-        super.onResume();
-    }
 
     private void showFormFragment() {
         fragmentTransaction = getSupportFragmentManager().beginTransaction();
@@ -77,11 +70,16 @@ public class MainActivity extends AppCompatActivity {
 
         initSearchMenu(menu);
 
-        ViewModelNote.getInstance().mutableLiveDataSelectedNoteList.observe(this, modelNotes -> {
-            menu.findItem(R.id.action_search).setVisible(modelNotes.isEmpty());
-            menu.findItem(R.id.action_deselect).setVisible(!modelNotes.isEmpty());
-            menu.findItem(R.id.action_all_select).setVisible(!ViewModelNote.getInstance().selectedCountEqualAllListSize());
-            menu.findItem(R.id.action_selected_delete).setVisible(!modelNotes.isEmpty());
+        ViewModelNote.getInstance().liveDataNotesSelected.observe(this, selectedSize -> {
+
+            ActionBar actionBar = getSupportActionBar();
+            if (actionBar != null) {
+                actionBar.setTitle(selectedSize == 0 ? getString(R.string.app_name) : selectedSize + " object selected");
+            }
+
+
+            menu.findItem(R.id.action_search).setVisible(selectedSize == 0);
+            menu.findItem(R.id.action_selected_delete).setVisible(selectedSize > 0);
         });
 
         return super.onCreateOptionsMenu(menu);
@@ -90,16 +88,11 @@ public class MainActivity extends AppCompatActivity {
     @SuppressLint("NonConstantResourceId")
     @Override
     public boolean onOptionsItemSelected(@NonNull MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.action_all_select:
-                ViewModelNote.getInstance().selectAll();
-                return true;
-            case R.id.action_selected_delete:
-                if (!ViewModelNote.getInstance().deleteSelected(this)) {
-                    AppUtils.showSnackBar(binding.frameLayoutMain, "Selected notes are not deleted");
-                }
-                return true;
-            default:
+        if (item.getItemId() == R.id.action_selected_delete) {
+            if (!ViewModelNote.getInstance().deleteSelected(this)) {
+                AppUtils.showSnackBar(binding.frameLayoutMain, "Selected notes are not deleted");
+            }
+            return true;
         }
         return super.onOptionsItemSelected(item);
     }
@@ -110,8 +103,6 @@ public class MainActivity extends AppCompatActivity {
 
 
         menu.findItem(R.id.action_search).setVisible(true);
-        menu.findItem(R.id.action_deselect).setVisible(false);
-        menu.findItem(R.id.action_all_select).setVisible(false);
         menu.findItem(R.id.action_selected_delete).setVisible(false);
 
         MenuItem menuItem = menu.findItem(R.id.action_search);
